Extract the handle typeahead into shared, reusable pieces
Pull the three parts of the Connectors card's handle typeahead into shared
homes so a second surface can reuse them behind the single
`atmosphere_handle_typeahead_url` filter:

- the endpoint resolver into `Atmosphere\handle_typeahead_url()`
  (`includes/functions.php`, alongside `appview_url()`);
- the `HandleTypeahead` component and its `searchHandles` helper into
  `src/shared/handle-typeahead.js`, taking the endpoint as a prop;
- the handle-field styles into `assets/css/handle-typeahead.css`.

The Connectors card consumes these shared pieces with no change to its
behavior.
